+++ acf-field-group-values.php
@@ -7,10 +7,10 @@
  * Author URI:      https://www.timjensen.us
  * Text Domain:     acf-field-group-values
  * Domain Path:     /languages
- * Version:         2.3.0
+ * Version:         3.0.0
  *
  * @package         TimJensen\ACF\Field_Group_Values
- * @license         GPL-3.0+
+ * @license         GPL-3.0-or-later
  */
 
 if ( ! defined( 'ABSPATH' ) ) {