+++ tests/fixtures/test_keys.php
@@ -1,0 +1,22 @@
+<?php
+
+/**
+ * Array keys are the meta keys in test_data.php. Array values are used for looking up return values.
+ */
+return [
+	'group_text'                                               => 'group.text',
+	'group_nestedgroup_subgroup1'                              => 'group.nestedgroup.subgroup1',
+	'group_nestedgroup_subgroup2'                              => 'group.nestedgroup.subgroup2',
+	'repeater_0_subfield'                                      => 'repeater.0.subfield',
+	'repeater_0_nestedrepeater_0_subfield'                     => 'repeater.0.nestedrepeater.0.subfield',
+	'repeater_0_nestedrepeater_1_subfield'                     => 'repeater.0.nestedrepeater.1.subfield',
+	'flexcontent_0_textfield'                                  => 'flexcontent.0.textfield',
+	'flexcontent_1_wysiwyg'                                    => 'flexcontent.1.wysiwyg',
+	'clone-text'                                               => 'clone-text',
+	'clone-repeater_0_subfield'                                => 'clone-repeater.0.subfield',
+	'clone-repeater_1_subfield'                                => 'clone-repeater.1.subfield',
+	'clone-flexcontent_0_textfield'                            => 'clone-flexcontent.0.textfield',
+	'clone-flexcontent_0_cloned-text-field-prefixed_textfield' => 'clone-flexcontent.0.cloned-text-field-prefixed_textfield',
+	'subfield'                                                 => 'subfield',
+];
+