+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log for ACF Field Group Values
 
+## 3.0.0 - 2018-07-21
+### Fixed
+- Changed array keys used for storing clone fields to be more consistent with how ACF stores the meta in the database. **This is a breaking change for those using clone fields.** 
+- Added support for prefixing clone fields
+- Added support for nested clone fields
+- Update tests to reflect the change related to storing clone field values 
+- Renamed test data directory
+
 ## 2.3.0 - 2018-05-06
 ### Added
 - Mutation testing using Infection
@@ -8,7 +16,7 @@
 - Support for 'option' or 'options' key
 
 ### Updated
-- Travis ci script
+- Travis CI script
 
 ### Fixed
 - Fatal error in rare cases when the package is installed as both a Composer dependency and a WordPress plugin, AND the site does not meet the PHP requirement.